pyproject.toml
[project]
name = "mllm_defake"
version = "0.5.0"
description = 'A flexible & scalable MLLM-based AIGC detection pipeline'
readme = {"file" = "readme.md", "content-type" = "text/markdown"}
requires-python = ">=3.10"
license = "Apache-2.0 OR MIT"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
]
dependencies = [
    "click>=8.0.1",
    "orjson",
    "pandas>=2",
    "numpy<2",
    "loguru",
    "tqdm",
    "openai",
    "requests",
    "scikit-learn>=0.21.0",
    "pillow",
    "tabulate",
    "opencv-python",
    "filetype",
    "torch>=2.0.0"  # to inherit Dataset class and used for inference
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "isort", "pytest-cov"]
finetune = [
    "deepspeed==0.16.9",
    "imagesize==1.4.1",
    "ms-swift==3.5.3",
    "qwen_vl_utils>=0.0.6",
    "torch==2.6.0",
    "vllm==0.8.5.post1"
]
comfor = ["timm>=0.9.12", "torchvision>=0.17.1"]

[project.scripts]
mllmdf = "mllm_defake.cli:main_cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
exclude = [
    "/.github",
    "/docs",
    "/dist",
]

[tool.hatch.build.targets.wheel]
packages = ["mllm_defake"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B"]
# ignore = ["E501"]
